feat(backend): add nutrition plan and user models
- Add NutritionPlan model for diet management
- Add GoalPlan model for fitness goals
- Add UserJwtInfo model for JWT authentication
- Add Gender and Role enums
- Add ChoosePlanResponseDto
- Update User and Profile models with new fields

+++ back-end/fitlife-planner-back-end/Api/Responses/CreateProfileResponse.cs
@@ -1,3 +1,4 @@
+using fitlife_planner_back_end.Api.Enums;
 using fitlife_planner_back_end.Api.Models;
 
 
@@ -6,7 +7,7 @@ namespace fitlife_planner_back_end.Api.Responses
     public class CreateProfileResponseDto
     {
         public CreateProfileResponseDto(Guid id, Guid userId, string displayName, string avatarUrl, DateTime birthDate,
-            Gender gender, float heightCm, float weightKg, float bmi, string bio, Dictionary<string, object> goals,
+            Gender gender, double heightCm, double weightKg, double bmi, string bio, Dictionary<string, object> goals,
             DateTime createAt, DateTime updateAt, int version)
         {
             Id = id;
@@ -32,9 +33,9 @@ namespace fitlife_planner_back_end.Api.Responses
         public string AvatarUrl { get; set; }
         public DateTime BirthDate { get; set; }
         public Gender Gender { get; set; }
-        public float HeightCm { get; set; }
-        public float WeightKg { get; set; }
-        public float BMI { get; set; }
+        public double HeightCm { get; set; }
+        public double WeightKg { get; set; }
+        public double BMI { get; set; }
         public string Bio { get; set; }
         public Dictionary<string, object> Goals { get; set; }
         public DateTime CreateAt { get; set; }
